[perf] Move alerts over to Cloud Datastore.
Also update UI for Alerts to show display name instead of id.

Bug: skia:

+++ perf/go/alerts/config_test.go
@@ -14,7 +14,7 @@ func TestConfig(t *testing.T) {
 	a := NewConfig()
 	assert.Equal(t, "-1", a.IdAsString())
 	a.StringToId("2")
-	assert.Equal(t, 2, a.ID)
+	assert.Equal(t, int64(2), a.ID)
 	assert.Equal(t, "2", a.IdAsString())
 }
